requirements.txt
greenlet==0.4.9
msgpack-python==0.4.6
wheel==0.24.0
django-flat-theme==1.1.1
simple-django-tequila==2.0.2

-r djangobb/requirements.txt
